removed unecessary fields from compile form and added polite message if backend is off

+++ app.js
@@ -38,15 +38,25 @@ app.use(bodyParser.urlencoded({extended: false}));
 app.post('/compilex', function(req, res) {
   // Specifies which URL to listen for
   // req.body -- contains form data
-  
+  //console.log("req.body.codesend='"+req.body.codesend+"'");
+
+ if(!process.env.DOCKERNEOCOMPILER) {
+  console.log("Error! No DOCKERNEOCOMPILER variable is set!\n");
+  var msg64 = new Buffer("Unable to communicate with backend compiler. Please try again later.",'ascii').toString('base64');
+  var msgret = "{\"output\":\""+msg64+"\",\"avm\":\"\",\"abi\":\"\"}";
+  //console.log("output is: '"+msgret+"'");
+  res.send(msgret);
+ }
+ else {
   var code64 = new Buffer(req.body.codesend, 'ascii').toString('base64');
   var cmddocker = "docker run -e COMPILECODE="+code64+" -t --rm $DOCKERNEOCOMPILER";
   var outp = "";
   outp = require('child_process').execSync(cmddocker).toString();
   //console.log("returning json..."+outp);
-  console.log("output is: '"+outp+"'");
+  //console.log("output is: '"+outp+"'");
   //res.send(JSON.stringify(outp));
   res.send(outp);
+ }
 });
 
 // catch 404 and forward to error handler
